requirements.txt
pyTelegramBotAPI==3.6.6
psycopg2-binary==2.8.4
yoyo_migrations==6.1.0